Refactor estructure
