+++ pom.xml
@@ -74,6 +74,11 @@
 			<artifactId>aws-java-sdk</artifactId>
 			<version>1.12.196</version>
 		</dependency>
+		<dependency>
+			<groupId>io.jsonwebtoken</groupId>
+			<artifactId>jjwt</artifactId>
+			<version>0.9.1</version>
+		</dependency>
 	</dependencies>
 
 	<build>
