+++ src/main/java/com/alkemy/ong/auth/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.security;
+package com.alkemy.ong.auth.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpMethod;
