+++ src/main/java/com/alkemy/ong/auth/controller/UserAuthController.java
@@ -1,19 +1,16 @@
-package com.alkemy.ong.controller;
+package com.alkemy.ong.auth.controller;
 
-import com.alkemy.ong.dto.UserProfileDto;
-import com.alkemy.ong.service.IUserAuthService;
+import com.alkemy.ong.auth.dto.UserProfileDto;
+import com.alkemy.ong.auth.service.IUserAuthService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.servlet.http.HttpServletRequest;
-
 @RestController
-@RequestMapping("/auth")
-public class UserAuthController {
 
+public class UserAuthController {
 
     @Autowired
     private IUserAuthService iUserAuthService;
